Ghost vector utilities added to chi_math::PETScUtils
