+++ CHI_TECH/ChiMath/PETScUtils/petsc_utils.h
@@ -24,11 +24,7 @@ namespace chi_math
       int    maximum_iterations = 100;
     };
 
-    Mat CreateSquareMatrix(int local_size, int global_size);
-    void CreateSquareMatrix(Mat& A, int local_size, int global_size);
-    void InitMatrixSparsity(Mat A,
-                            std::vector<int>& nodal_nnz_in_diag,
-                            std::vector<int>& nodal_nnz_off_diag);
+    //01
     Vec CreateVector(int local_size, int global_size);
     void CreateVector(Vec& x, int local_size, int global_size);
 
@@ -36,6 +32,14 @@ namespace chi_math
                                int nghosts,
                                std::vector<int>& ghost_indices);
 
+    //02
+    Mat CreateSquareMatrix(int local_size, int global_size);
+    void CreateSquareMatrix(Mat& A, int local_size, int global_size);
+    void InitMatrixSparsity(Mat A,
+                            std::vector<int>& nodal_nnz_in_diag,
+                            std::vector<int>& nodal_nnz_off_diag);
+
+    //03
     PETScSolverSetup CreateCommonKrylovSolverSetup(
       Mat ref_matrix,
       std::string in_solver_name = "KSPSolver",
@@ -54,12 +58,34 @@ namespace chi_math
     GeneralKSPMonitor(KSP ksp, PetscInt n,
                       PetscReal rnorm, void *monitordestroy);
 
+    //04
     void CopyVecToSTLvector(Vec x, std::vector<double>& data, size_t N);
 
     void CopyGlobalVecToSTLvector(
       Vec x,
       const std::vector<int>& global_indices,
       std::vector<double>& data);
+
+    void CommunicateGhostEntries(Vec x);
+
+    /**Simple data structure to keep track of a ghost vector's
+     * localized views.*/
+    struct GhostVecLocalRaw
+    {
+      Vec     x_localized;
+      double* x_localized_raw;
+
+      /**Returns a copy of the value at the specified index.*/
+      double operator[](int index)
+      {return x_localized_raw[index];}
+
+      /**Returns a reference of the value at the specified index.*/
+      double& operator()(int index)
+      {return x_localized_raw[index];}
+    };
+
+    GhostVecLocalRaw GetGhostVectorLocalViewRead(Vec x);
+    void RestoreGhostVectorLocalViewRead(Vec x,GhostVecLocalRaw& local_data);
   }
 }
